Secure the route

+++ app/api/issues/[id]/route.ts
@@ -1,11 +1,16 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { issueScheme } from '@/app/validationSchema';
 import prisma from '@/prisma/client';
+import { getServerSession } from 'next-auth';
+import authOptions from '@/app/auth/authOptions';
 
 export async function PATCH(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  const session = await getServerSession(authOptions);
+  if (!session) return NextResponse.json({}, { status: 401 });
+
   const body = await request.json();
   const validation = issueScheme.safeParse(body);
 
@@ -33,23 +38,23 @@ export async function PATCH(
 }
 
 export async function DELETE(
-   request: NextRequest,
-   { params }: { params: { id: string } }
- ) {
-  
- 
-   const issue = await prisma.issue.findUnique({
-     where: { id: parseInt(params.id) },
-   });
- 
-   if (!issue) {
-     return NextResponse.json({ error: 'Invalid issue' }, { status: 404 });
-   }
- 
-   await prisma.issue.delete({
-      where: { id: issue.id }
-    });
-  
-    return NextResponse.json({});
- }
- 
+  request: NextRequest,
+  { params }: { params: { id: string } }
+) {
+  const session = await getServerSession(authOptions);
+  if (!session) return NextResponse.json({}, { status: 401 });
+
+  const issue = await prisma.issue.findUnique({
+    where: { id: parseInt(params.id) },
+  });
+
+  if (!issue) {
+    return NextResponse.json({ error: 'Invalid issue' }, { status: 404 });
+  }
+
+  await prisma.issue.delete({
+    where: { id: issue.id },
+  });
+
+  return NextResponse.json({});
+}
